+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ### 10.0 [not yet released]
 
-- replaced IntsRef with BytesRef. Note that setInt(edgeId, index) uses a byte index now, i.e. instead of edgeAccess.setInt(edgeId, index, value) you now have to use edgeAccess.setInt(edgeId, index * 4, value)
-
+- replaced IntsRef with BytesRef and EdgeIntAccess with EdgeBytesAccess. Note that EdgeBytesAccess.setInt(edgeId, index) requires a byte index!
 
 ### 9.0 [23 Apr 2024]